# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SWC Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/": {},
        "/api/events": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Get all Events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "First Event"
                                                    },
                                                    "descr": {
                                                        "type": "string",
                                                        "example": "First Event Description"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Event"
                ],
                "summary": "Create Event",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "user_id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "descr": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object",
                                        "example": {
                                            "user_id": 1,
                                            "title": "First Event",
                                            "descr": "First Event Description"
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "First Event"
                                                },
                                                "descr": {
                                                    "type": "string",
                                                    "example": "First Event Description"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{event}": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Get single Event by id",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID event",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "First Event"
                                                    },
                                                    "descr": {
                                                        "type": "string",
                                                        "example": "First Event Description"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Event"
                ],
                "summary": "Delete Event by id",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID event",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Event"
                ],
                "summary": "Patch Event by id",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID event",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "anyOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "descr": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object",
                                        "example": {
                                            "title": "First Event",
                                            "descr": "First Event Description"
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "First Event"
                                                    },
                                                    "descr": {
                                                        "type": "string",
                                                        "example": "First Event Description"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "500": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "result": {
                                    "type": "object"
                                },
                                "error": {
                                    "type": "string",
                                    "example": "Error"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "apiKey",
                "description": "!!!! ENTER TOKEN FORMAT with Bearer(Bearer token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "apiAuth": []
        }
    ]
}